Install hyper-v integration services
The virtualbox output will have both virtualbox and hyper-v integration tools installed.

This registry entry will trigger windows update to offer KB3063283 and KB3063109, which
are required for the VM to work in Hyper-V on a Windows 10 or Windows 2016 host

devcon.exe is only available in the Windows Driver kit (a 600 MB download) so it gets
commited instead. Using the 64-bit executable.